services/ingestion-worker/requirements.txt
redis>=5.0.0
supabase>=2.10.0
python-dotenv>=1.0.0
neo4j>=5.0.0
lsprotocol>=2023.0.0
